requirements.txt
pillow>=9.0.0
python-decouple
python-telegram-bot[callback-data]>=22.0
